+++ src/main/scala/org/apache/spark/ml/feature/VectorAccumulator.scala
@@ -7,22 +7,22 @@ import org.apache.spark.util.AccumulatorV2
 /**
  * @author sramirez
  */
-class VectorAccumulator(val rows: Int) extends AccumulatorV2[Vector[Float], Vector[Float]] {
+class VectorAccumulator(val rows: Int) extends AccumulatorV2[Vector[Double], Vector[Double]] {
   
-  def this(m: Vector[Float]) = {
+  def this(m: Vector[Double]) = {
     this(m.size)
     this.accVector = m.copy
   }
 
-  private var accVector: Vector[Float] = Vector.zeros[Float](rows)
+  private var accVector: Vector[Double] = Vector.zeros[Double](rows)
   private var zero: Boolean = true
 
   def reset(): Unit = {
-    accVector = Vector.zeros[Float](rows)
+    accVector = Vector.zeros[Double](rows)
     zero = true
   }
 
-  def add(v: Vector[Float]): Unit = {
+  def add(v: Vector[Double]): Unit = {
     if(isZero) 
       zero = false
     accVector += v
@@ -30,9 +30,9 @@ class VectorAccumulator(val rows: Int) extends AccumulatorV2[Vector[Float], Vect
   
   def isZero(): Boolean = zero
   
-  def merge(other: AccumulatorV2[Vector[Float], Vector[Float]]): Unit = accVector += other.value
+  def merge(other: AccumulatorV2[Vector[Double], Vector[Double]]): Unit = accVector += other.value
   
-  def value: Vector[Float] = accVector
+  def value: Vector[Double] = accVector
   
-  def copy(): AccumulatorV2[Vector[Float], Vector[Float]] = new VectorAccumulator(accVector)
+  def copy(): AccumulatorV2[Vector[Double], Vector[Double]] = new VectorAccumulator(accVector)
 }
